Automatic hostel room allotment system.
: It is an command based system in which student fill some information
like name type of room he need like 1st seated, 2nd seated etc and his
course and year. And admin allot room in such a way that all the
students in a room are of same year and have only that type of room he
is applied for.  Admin can also cancel the allotment of any student at
any time and allot other student from waiting list of same year and
request for same type of room.  All allotment is based on first come and
first serve manner.